4 task + design + add full info about card

+++ app/src/main/java/com/example/shift_tz_2023/data/database/entities/CountryEntity.kt
@@ -10,8 +10,8 @@ data class CountryEntity(
     val alpha2: String,
     val currency: String,
     val emoji: String,
-    val latitude: Int,
-    val longitude: Int,
+    val latitude: Double,
+    val longitude: Double,
     val name: String,
     val numeric: String
 )
